chore: update dependencies and configurations for legacy Granite HF / SSM environment
- Updated `torch` versioning in `pyproject.toml`, `requirements.txt`, and legacy requirement files to differentiate between Linux and non-Linux platforms.
- Added additional dependencies (`pandas`, `ollama`, `opencc`, `openai`, `pyyaml`, `tqdm`) to legacy requirement files.
- Changed default `torch_dtype` from `bfloat16` to `float16` in model configuration and script arguments for consistency.

+++ scripts/smoke_vllm_offline_model.py
@@ -29,7 +29,7 @@ def build_parser() -> argparse.ArgumentParser:
     parser.add_argument("--prompt", default=DEFAULT_PROMPT)
     parser.add_argument("--trust-remote-code", dest="trust_remote_code", action="store_true")
     parser.add_argument("--no-trust-remote-code", dest="trust_remote_code", action="store_false")
-    parser.add_argument("--torch-dtype", default="bfloat16")
+    parser.add_argument("--torch-dtype", default="float16")
     parser.add_argument("--gpu-memory-utilization", type=float, default=0.9)
     parser.add_argument("--max-model-len", type=int, default=131072)
     parser.add_argument("--max-tokens", type=int, default=48)
